requirements.txt
streamlit==1.32.2
openai==1.12.0
pinecone-client==3.0.0
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.6.0